pyproject.toml
[project]
name = "RDCVentilationControl"
version = "0.0.1"
description = "RDC's ventilation control project."
authors = [{ name = "ashkf", email = "ifan.kai@outlook.com" }]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10.11"
dependencies = [
    "python=3.10.11",
    "torch=2.1.0",
    "pip=23.2.1",
    "matplotlib=3.7.1",
    "numpy=1.25.0",
    "pandas=1.5.3",
    "Pillow=9.4.0",
    "PyYAML=6.0.0",
    "Requests=2.29.0",
    "scikit-learn=1.2.2",
    "torchmetrics=0.11.4",
    "tqdm=4.65.0",
    "torchinfo=1.8.0",
    "cnocr=2.2.3.2",
    "pyDOE2=1.3.0",
    "scienceplots=2.0.1",
    "torchview=0.2.6",
    "geatpy=2.7.0",
]

# https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[tool.black]
line-length = 88
preview = true

# https://beta.ruff.rs/docs/settings/
[tool.ruff]
line-length = 88
task-tags = ["TODO", "todo", "FIXME", "fixme", "BUG", "bug", "NOTE", "note"]
# https://beta.ruff.rs/docs/rules/
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line-length, handled by black
    # "B008", # do not perform function calls in argument defaults
    # "C901", # too complex
]
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
ignore-init-module-imports = true

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
known-first-party = ["REPLACE_PACKAGE_NAME"]
